+++ src/util/createSkeletonData.js
@@ -1,6 +1,9 @@
+import MovieItem from '../component/common/MovieItem'
+
 const createSkeletonData = Array(20).fill({
   poster_path: null,
   title: null,
   vote_average: null,
 })
+
 export default createSkeletonData
